requirements.txt
edge-tts==7.0.1
moviepy==1.0.3
numpy==2.2.4
pillow==9.5.0
torch==2.6.0
torchaudio==2.6.0
whisperx @ git+https://github.com/m-bain/whisperx.git@0aed8745890f12ecfe0b2d9c4ba62bcdfb16f94e
